refactor: renames ownerUtils into owner
