+++ frontend/amundsen_application/static/js/utils/owner.ts
